% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allcellcb.R
\name{allcellcb}
\alias{allcellcb}
\title{allcellcb}
\usage{
allcellcb(cellsets)
}
\arguments{
\item{cellsets}{A vector containing the names of the cell types(length(cellsets)>=3) of interest.}
}
\value{
A list containing all combinations of cell types
}
\description{
Lists all combinations of cell types
}
\author{
W Xie
}
