% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkinput.R
\name{checkinput}
\alias{checkinput}
\title{checkinput}
\usage{
checkinput(metadata, cdata, cellsets)
}
\arguments{
\item{metadata}{A dataframe containing the 'sample' and 'condition' columns.
The class of "condition" should be factor, and levels of "condition" should be "2".}

\item{cdata}{A dataframe with row names for sample names and column names containing at least three cell type names.}

\item{cellsets}{A vector containing the names of the cell types(length(cellsets)>=3) of interest.}
}
\description{
What this function does is check:
\enumerate{
\item Whether the class of each input data is correct
\item whether the sample names of "metadata" and "cdata" are the same.
\item whether the class of the "condition" is factor, and there are only two levels;
\item Whether length(cellsets)>=3
If the above conditions are met, the execution continues, otherwise, it will be stopped
}
}
\details{
Check input data
}
\author{
W Xie
}
