% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commroc.R
\name{commroc}
\alias{commroc}
\title{commroc}
\usage{
commroc(
  cdata,
  metadata,
  cellsets,
  sami = 500,
  ncores = (parallel::detectCores() - 1)
)
}
\arguments{
\item{cdata}{A dataframe with row names for sample names
and column names containing at least three cell type names.}

\item{metadata}{A dataframe containing the 'sample' and 'condition' columns.
The class of "condition" should be factor, and levels of "condition" should be "2".}

\item{cellsets}{A vector containing the names of the cell types(length(cellsets)>=3) of interest.}

\item{sami}{A integer repersenting the times of random sampling, defualt value is 500.}

\item{ncores}{the number of CPU cores to use,
for Linux systerm,default value is detected the number of CPU cores on the current host - 1}
}
\value{
a dataframe, which one column is roc.auc(mean of ROC value for all combinations)
and the other column is cellsets(cell types).
}
\description{
The function calculates all combinations of the cell types of interest and then the average AUC value for the times of user-specified sampling under each combination
}
\details{
Calulate average AUC values of specified times of sampling for all combinations
}
\examples{
set.seed(1)
cdata <- matrix(sample(100:300, 100, replace = TRUE), 20, 5)
row.names(cdata) <- paste0("A",1:20)
colnames(cdata) <- c("B.cells.naive", "B.cells.memory", "Plasma.cells",
"T.cells.CD8", "T.cells.CD4.naive")
cdata <- as.data.frame(cdata)
set.seed(1)
metadata <- data.frame(sample = row.names(cdata),
condition = sample(c("CR/PR", "SD/PD"), 20, replace = TRUE))
metadata$condition <- as.factor(metadata$condition)
cellsets <- c("B.cells.naive", "B.cells.memory", "Plasma.cells", "T.cells.CD8")
sami = 5
commroc(cdata = cdata, metadata = metadata, cellsets = cellsets, sami = sami, ncores = 2)
}
\author{
W Xie
}
