% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divcells.R
\name{divcells}
\alias{divcells}
\title{divcells}
\usage{
divcells(cdata, metadata, slcells)
}
\arguments{
\item{cdata}{A dataframe with row names for sample names and column names containing at least three cell type names.}

\item{metadata}{A dataframe containing the 'sample' and 'condition' columns.
The class of "condition" should be factor, and levels of "condition" should be "2".}

\item{slcells}{A vector containing selected cell types}
}
\value{
a dataframe which the row names are the sample names,
and the column names are sample (sample name),
divscore (diversity score of the selected cell type),
and condition respectively.
}
\description{
Calculate diversity scores(with metadata)
}
\author{
W Xie
}
