% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commroc.R
\name{maxroc}
\alias{maxroc}
\title{maxroc}
\usage{
maxroc(acroc)
}
\arguments{
\item{acroc}{a dataframe, which one column is roc.auc(mean of ROC value for all combinations)
and the other column is cellsets(cell types)}
}
\description{
Selecte the combination of cell types corresponding to the maximum AUC value
}
\examples{
acroc <- data.frame(roc.auc = c(0.440625, 0.493750, 0.690625),
  cellsets = c("B.cells.naive; B.cells.memory; Plasma.cells",
    "B.cells.naive; B.cells.memory; T.cells.CD8",
    "B.cells.naive; Plasma.cells; T.cells.CD8"))
maxroc(acroc = acroc)
}
\author{
W Xie
}
