% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meansamroc.R
\name{meansamroc}
\alias{meansamroc}
\title{meansamroc}
\usage{
meansamroc(sami, samsize, samp1, samp2, scoreif)
}
\arguments{
\item{sami}{the times of random sampling}

\item{samsize}{a vector containing two numeric values
representing the number of samples for the two types of conditions.}

\item{samp1}{a vector containing all sample names in the first condition}

\item{samp2}{a vector containing all sample names in the second condition}

\item{scoreif}{a dataframe which the row names are the sample names,
and the column names are sample (sample name),
divscore (diversity score of the selected cell type), and condition respectively.}
}
\value{
a numeric representing the average AUC value for the specified times of sampling
}
\description{
Calculate The average AUC value for the specified times of sampling
}
\author{
W Xie
}
