% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProp.R
\name{plot.flow}
\alias{plot.flow}
\title{plot.flow}
\usage{
\method{plot}{flow}(dat = dat, xtitle = xtitle, et = et, bar_width = bar_width)
}
\arguments{
\item{dat}{a dataframe containing three columns: xtitle(self-defined), CellType, Proportion}

\item{xtitle}{A custom column name, which is also the title name of the X-axis}

\item{et}{element_text}

\item{bar_width}{the width of the bar}
}
\value{
a bar chart
}
\description{
Alluvial Plots
}
\author{
W Xie
}
