% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCC.R
\name{plotCC}
\alias{plotCC}
\title{plotCC}
\usage{
plotCC(
  data,
  x = "condition",
  y = "divscore",
  xlabs = "Condition",
  ylabs = "DiversityScore",
  method = "wilcox.test",
  paired = FALSE,
  method.args = list(),
  ref.group = NULL,
  comparisons = NULL,
  hide.ns = FALSE,
  label.x = 1.5,
  label.y = 2,
  label = "p.format",
  color = NULL
)
}
\arguments{
\item{data}{A dataframe containing three coloums:"sample", "divscore", "condition"}

\item{x}{A "character"}

\item{y}{A "character"}

\item{xlabs}{xlabs}

\item{ylabs}{ylabs}

\item{method}{a character string indicating which method to be used for
comparing means.}

\item{paired}{a logical indicating whether you want a paired test. Used only
in \code{\link[stats]{t.test}} and in \link[stats]{wilcox.test}.}

\item{method.args}{a list of additional arguments used for the test method.
For example one might use \code{method.args = list(alternative = "greater")}
for wilcoxon test.}

\item{ref.group}{a character string specifying the reference group. If
 specified, for a given grouping variable, each of the group levels will be
 compared to the reference group (i.e. control group).

 \code{ref.group} can be also \code{".all."}. In this case, each of the
 grouping variable levels is compared to all (i.e. basemean).}

\item{comparisons}{A list of length-2 vectors. The entries in the vector are
either the names of 2 values on the x-axis or the 2 integers that correspond
to the index of the groups of interest, to be compared.}

\item{hide.ns}{logical value. If TRUE, hide ns symbol when displaying
significance levels.}

\item{label.x, label.y}{\code{numeric} Coordinates (in data units) to be used
for absolute positioning of the label. If too short they will be recycled.}

\item{label}{character string specifying label type. Allowed values include
"p.signif" (shows the significance levels), "p.format" (shows the formatted
p value).}

\item{color}{a vector containing a set of aesthetic values}
}
\value{
a plot of the difference of diversity scores
}
\description{
Plot the diversity score comparison between the two conditions
}
\examples{
set.seed(1)
data1 <- data.frame(sample = c(paste0("A", 1:50)),
divscore = runif(50, min = 0, max = 0.7),
condition = rep("condition1", 50))
set.seed(1)
data2 <- data.frame(sample = c(paste0("A", 1:50)),
divscore = runif(50, min = 0.7, max = 1.3),
condition = rep("condition2", 50))
data = rbind(data1, data2)
plotCC(data = data, x = "condition", y = "divscore",
xlabs = "Condition", ylabs = "DiversityScore",
color = c("#00AED7", "#FD9347"),
label.x = 1.5, label.y = 1.7,
comparisons = NULL)
}
\author{
W Xie
}
