% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProp.R
\name{plotProp}
\alias{plotProp}
\title{plotProp}
\usage{
plotProp(
  cdata,
  slcells = colnames(cdata),
  xtitle = "Sample",
  axistext.x = TRUE,
  bar_width = 0.5,
  flow = TRUE,
  geomtext = TRUE,
  color = NULL
)
}
\arguments{
\item{cdata}{a dataframe or matrix with row names
for sample names and column names containing at least three cell type names}

\item{slcells}{selected cell types}

\item{xtitle}{a character representing the title of x axis}

\item{axistext.x}{a Boolean value that represents whether x-axis text is displayed}

\item{bar_width}{the width of the bar}

\item{flow}{a Boolean value representing whether it is presented as Alluvial Plots,
default value is TRUE}

\item{geomtext}{a Boolean value representing whether Annotate value,
default value is TRUE}

\item{color}{a vector containing a set of aesthetic values}
}
\value{
a bar chart representing cell types proportion
}
\description{
plot selected cell types proportion
}
\examples{
set.seed(1)
cdata <- matrix(sample(100:300, 25, replace = TRUE), 5, 5)
row.names(cdata) <- paste0("A",1:5)
colnames(cdata) <- c("B.cells.naive", "B.cells.memory", "Plasma.cells",
"T.cells.CD8", "T.cells.CD4.naive")
slcells <- c("B.cells.naive", "B.cells.memory", "Plasma.cells")
plotProp(cdata = cdata, slcells = slcells, xtitle = "Sample",
axistext.x = TRUE, bar_width = 0.5, flow = TRUE, geomtext = TRUE,
color = c("#00468BFF", "#925E9FFF", "#AD002AFF", "#FDAF91FF"))
}
\author{
W Xie
}
