% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocdir.R
\name{rocdir}
\alias{rocdir}
\title{rocdir}
\usage{
rocdir(scoreif)
}
\arguments{
\item{scoreif}{A dataframe which the row names are the sample names,
and the column names are sample (sample name),
divscore (diversity score of the selected cell type), and condition respectively.}
}
\value{
direction symbol: "<" or ">"
}
\description{
Determine the direction of ROC by using all samples
}
\author{
W Xie
}
