% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sami_reset.R
\name{sami_reset}
\alias{sami_reset}
\title{sami reset}
\usage{
sami_reset(metadata, sami)
}
\arguments{
\item{metadata}{A dataframe containing the 'sample' and 'condition' columns.
The class of "condition" should be factor, and levels of "condition" should be "2".}

\item{sami}{A numeric representing the times of random sampling}
}
\value{
A numeric representing the times of random sampling taken after the reset
}
\description{
Reset the sampling times forcibly if the data is balanced
}
\author{
W Xie
}
