% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saminfor.R
\name{sampcondi}
\alias{sampcondi}
\title{sampcondi}
\usage{
sampcondi(metadata, condi = 1)
}
\arguments{
\item{metadata}{A dataframe containing the 'sample' and 'condition' columns.
The class of "condition" should be factor, and levels of "condition" should be "2".}

\item{condi}{A numeric, which can be 1 or 2, representing the first condition and the second condition respectively. The default value is 1}
}
\value{
a vector containing all sample names in the specified condition
}
\description{
Lists the sample names contained in the specified condition
}
\author{
W Xie
}
