% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saminfor.R
\name{samsizes}
\alias{samsizes}
\title{samsizes}
\usage{
samsizes(metadata)
}
\arguments{
\item{metadata}{A dataframe containing the 'sample' and 'condition' columns.
The class of "condition" should be factor, and levels of "condition" should be "2".}
}
\value{
A vector containing two numeric values representing the number of samples for the two types of conditions.
}
\description{
The samples of the two conditions were counted respectively
}
\author{
W Xie
}
